requirements.txt
fastapi
uvicorn[standard]
requests
pillow
torch
torchvision
sentence-transformers
psycopg2-binary
python-dotenv
sqlalchemy
